requirements.txt
torch>=2.0.0
torchtext>=0.15.0
numpy>=1.24.0
sentencepiece>=0.1.99
sacrebleu>=2.3.0
pyyaml>=6.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.12.0
fastapi>=0.95.0
uvicorn[standard]>=0.21.0
gradio>=3.35.0
streamlit>=1.22.0
python-multipart>=0.0.6
aiofiles>=23.1.0
pytest>=7.3.0
black>=23.3.0
flake8>=6.0.0
